Adjust Local filesystem adapter to not follow symlinks when deleting

+++ library/CM/File/Filesystem/Adapter/Local.php
@@ -56,7 +56,7 @@ class CM_File_Filesystem_Adapter_Local extends CM_File_Filesystem_Adapter implem
         if (!$this->exists($path)) {
             return;
         }
-        if ($this->isDirectory($path)) {
+        if ($this->isDirectory($path) && !$this->_isLink($path)) {
             if (false === @rmdir($pathAbsolute)) {
                 throw new CM_Exception('Cannot delete directory `' . $pathAbsolute . '`');
             }
@@ -149,6 +149,9 @@ class CM_File_Filesystem_Adapter_Local extends CM_File_Filesystem_Adapter implem
      * @throws CM_Exception
      */
     private function _listByPrefixRecursive($pathPrefix, array &$fileList, array &$dirList) {
+        if ($this->_isLink($pathPrefix)) {
+            return;
+        }
         $pathPrefixAbsolute = $this->_getAbsolutePath($pathPrefix);
         $filenameList = @scandir($pathPrefixAbsolute);
         if (false === $filenameList) {
@@ -168,6 +171,11 @@ class CM_File_Filesystem_Adapter_Local extends CM_File_Filesystem_Adapter implem
         foreach ($fileListLocal as $filePath) {
             $fileList[] = $filePath;
         }
+    }
+
+    private function _isLink($path) {
+        $pathAbsolute = $this->_getAbsolutePath($path);
+        return is_link($pathAbsolute);
     }
 
     /**
